% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flx_mov_desc.R
\name{flx_mov_desc}
\alias{flx_mov_desc}
\title{Get the description of a movie}
\usage{
flx_mov_desc(movie = NULL)
}
\arguments{
\item{movie}{The name of a movie as a character string or an XML document representing the HTML of of a movie page}
}
\value{
A character string of the description of the movie.
}
\description{
Get the description of a movie
}
\examples{
flx_mov_desc(movie = "The Godfather")
}
