% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flx_mov_details.R
\name{flx_mov_details}
\alias{flx_mov_details}
\title{Get a variety of movie details}
\usage{
flx_mov_details(movie)
}
\arguments{
\item{movie}{The name of a movie as a character string or an XML document representing the HTML of of a movie page}
}
\value{
A tibble with 8 columns and 1 row containing the movie name, description, genre, director, lead cast, runtime, user rating, and Metacritic rating.
}
\description{
Get a variety of movie details
}
\examples{
flx_mov_details(movie = "The Godfather")
}
