% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flx_mov_director.R
\name{flx_mov_director}
\alias{flx_mov_director}
\title{Get the director of a movie}
\usage{
flx_mov_director(movie = NULL)
}
\arguments{
\item{movie}{The name of a movie as a character string or an XML document representing the HTML of of a movie page}
}
\value{
A character vector of length one. The director of the movie.
}
\description{
Get the director of a movie
}
\examples{
flx_mov_director(movie = "The Godfather")
}
