% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flx_mov_html.R
\name{flx_mov_html}
\alias{flx_mov_html}
\title{Get HTML for a movie page}
\usage{
flx_mov_html(movie = NULL)
}
\arguments{
\item{movie}{The name of a movie as a character string.}
}
\value{
An XML document representing the HTML of the movie page.
}
\description{
Get HTML for a movie page
}
\examples{
flx_mov_html(movie = "The Godfather")
}
