% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flx_mov_lead_cast.R
\name{flx_mov_lead_cast}
\alias{flx_mov_lead_cast}
\title{Get the lead cast of a movie}
\usage{
flx_mov_lead_cast(movie = NULL)
}
\arguments{
\item{movie}{The name of a movie as a character string or an XML document representing the HTML of of a movie page}
}
\value{
A character vector. The lead cast of the movie.
}
\description{
Get the lead cast of a movie
}
\examples{
flx_mov_lead_cast(movie = "The Godfather")
}
