% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flx_mov_mc_rtg.R
\name{flx_mov_mc_rtg}
\alias{flx_mov_mc_rtg}
\title{Get the Metacritic rating of a movie}
\usage{
flx_mov_mc_rtg(movie = NULL)
}
\arguments{
\item{movie}{The name of a movie as a character string or an XML document representing the HTML of of a movie page}
}
\value{
A numeric vector of length one. The Metacritic rating of the movie.
}
\description{
Get the Metacritic rating of a movie
}
\examples{
flx_mov_mc_rtg(movie = "The Godfather")
}
