% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flx_mov_name.R
\name{flx_mov_name}
\alias{flx_mov_name}
\title{Get the movie name matched to the input character string}
\usage{
flx_mov_name(movie = NULL)
}
\arguments{
\item{movie}{The name of a movie as a character string.}
}
\value{
A character string of the movie matched to the input.
}
\description{
Get the movie name matched to the input character string
}
\examples{
flx_mov_name(movie = "The Godfather")
}
