% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flx_mov_runtime.R
\name{flx_mov_runtime}
\alias{flx_mov_runtime}
\title{Get the runtime of a movie}
\usage{
flx_mov_runtime(movie = NULL)
}
\arguments{
\item{movie}{The name of a movie as a character string or an XML document representing the HTML of of a movie page}
}
\value{
A numeric vector of length one. The number of minutes.
}
\description{
Get the runtime of a movie
}
\examples{
flx_mov_runtime(movie = "The Godfather")
}
